+++ src/institutions.rs
@@ -1,1 +1,7 @@
+pub use client::*;
+pub use model::*;
 
+mod client;
+mod model;
+
+// TODO: Unit tests